Replaced float32 with float64

+++ api-service/database/database.go
@@ -96,19 +96,19 @@ type MongoDatabaseInterface interface {
 	// GetDatabaseArchivelogStatusStats return a array containing the number of databases per archivelog status
 	GetDatabaseArchivelogStatusStats(location string, environment string, olderThan time.Time) ([]interface{}, utils.AdvancedErrorInterface)
 	// GetTotalDatabaseWorkStats return the total work of databases
-	GetTotalDatabaseWorkStats(location string, environment string, olderThan time.Time) (float32, utils.AdvancedErrorInterface)
+	GetTotalDatabaseWorkStats(location string, environment string, olderThan time.Time) (float64, utils.AdvancedErrorInterface)
 	// GetTotalDatabaseMemorySizeStats return the total of memory size of databases
-	GetTotalDatabaseMemorySizeStats(location string, environment string, olderThan time.Time) (float32, utils.AdvancedErrorInterface)
+	GetTotalDatabaseMemorySizeStats(location string, environment string, olderThan time.Time) (float64, utils.AdvancedErrorInterface)
 	// GetTotalDatabaseDatafileSizeStats return the total size of datafiles of databases
-	GetTotalDatabaseDatafileSizeStats(location string, environment string, olderThan time.Time) (float32, utils.AdvancedErrorInterface)
+	GetTotalDatabaseDatafileSizeStats(location string, environment string, olderThan time.Time) (float64, utils.AdvancedErrorInterface)
 	// GetTotalDatabaseSegmentSizeStats return the total size of segments of databases
-	GetTotalDatabaseSegmentSizeStats(location string, environment string, olderThan time.Time) (float32, utils.AdvancedErrorInterface)
+	GetTotalDatabaseSegmentSizeStats(location string, environment string, olderThan time.Time) (float64, utils.AdvancedErrorInterface)
 	// GetTotalExadataMemorySizeStats return the total size of memory of exadata
-	GetTotalExadataMemorySizeStats(location string, environment string, olderThan time.Time) (float32, utils.AdvancedErrorInterface)
+	GetTotalExadataMemorySizeStats(location string, environment string, olderThan time.Time) (float64, utils.AdvancedErrorInterface)
 	// GetTotalExadataCPUStats return the total cpu of exadata
 	GetTotalExadataCPUStats(location string, environment string, olderThan time.Time) (interface{}, utils.AdvancedErrorInterface)
 	// GetAverageExadataStorageUsageStats return the average usage of cell disks of exadata
-	GetAverageExadataStorageUsageStats(location string, environment string, olderThan time.Time) (float32, utils.AdvancedErrorInterface)
+	GetAverageExadataStorageUsageStats(location string, environment string, olderThan time.Time) (float64, utils.AdvancedErrorInterface)
 	// GetExadataStorageErrorCountStatusStats return a array containing the number of cell disks of exadata per error count status
 	GetExadataStorageErrorCountStatusStats(location string, environment string, olderThan time.Time) ([]interface{}, utils.AdvancedErrorInterface)
 	// GetExadataPatchStatusStats return a array containing the number of exadata per patch status
@@ -117,7 +117,7 @@ type MongoDatabaseInterface interface {
 	// SetLicenseCount set the count of a certain license
 	SetLicenseCount(name string, count int) utils.AdvancedErrorInterface
 	// SetLicenseCostPerProcessor set the cost per processor of a certain license
-	SetLicenseCostPerProcessor(name string, costPerProcessor float32) utils.AdvancedErrorInterface
+	SetLicenseCostPerProcessor(name string, costPerProcessor float64) utils.AdvancedErrorInterface
 	// SetLicenseUnlimitedStatus set the unlimited status of a certain license
 	SetLicenseUnlimitedStatus(name string, unlimitedStatus bool) utils.AdvancedErrorInterface
 
@@ -139,7 +139,7 @@ type MongoDatabaseInterface interface {
 	// ExistHostdata return true if the host specified by hostname exist, otherwise false
 	ExistHostdata(hostname string) (bool, utils.AdvancedErrorInterface)
 	// GetTechnologiesUsage return a map that contains the number of usages for every features
-	GetTechnologiesUsage(location string, environment string, olderThan time.Time) (map[string]float32, utils.AdvancedErrorInterface)
+	GetTechnologiesUsage(location string, environment string, olderThan time.Time) (map[string]float64, utils.AdvancedErrorInterface)
 }
 
 // MongoDatabase is a implementation
@@ -193,10 +193,10 @@ type SearchHostsFilters struct {
 	PhysicalHost                  string
 	OperatingSystem               string
 	Kernel                        string
-	LTEMemoryTotal                float32
-	GTEMemoryTotal                float32
-	LTESwapTotal                  float32
-	GTESwapTotal                  float32
+	LTEMemoryTotal                float64
+	GTEMemoryTotal                float64
+	LTESwapTotal                  float64
+	GTESwapTotal                  float64
 	IsMemberOfCluster             *bool
 	CPUModel                      string
 	LTECPUCores                   int
